Major Changes: Defines, Logging Pt. 2 	* Removed All includes all files with few exceptions 		* Platforms & Common accumulates the includes 		* Added nostd_*.h to be the core mechanism to 		  include for platforms and new modules. 		* The Structures is as follows: 			- Defines + Basic Types (platform.h) 			- Types & Global Variables 			- Core Structures (s8, marena, etc.) 		* This is followed for common, linux and win32. 	* Moved include/platform/common/log.h code into defines.h 		* This is so core functions like memcpy can assert 	* Created type.h to put all types for the base structures           to use. This includes Memory Union and Log Declarations.
This is important here: I am slightly moving away from "modularity"
into "framework-like". It wasn't my intention, however logging has
made it quite complicated. I may want to change this in the future,
as I like certain files to be "Standalone". Perhaps I even make a
standalone branch in where you can just grab a file (like s8) and
just use it. An attempt to exclude define.h and platform.h will be
tried but if it looks extremely ugly then the only dependency
would be define.h and platform.h. Perhaps this will be a problem
for future (and more experienced) me.